clickevent-ingest/requirements.txt
asyncio-nats-client==0.10.0
